Fix tests breaking
