+++ parachain/pallets/ethereum-beacon-client/Cargo.toml
@@ -19,7 +19,7 @@ ssz-rs = { git = "https://github.com/ralexstokes/ssz-rs", default-features = fal
 ssz-rs-derive = { git = "https://github.com/ralexstokes/ssz-rs", default-features = false, rev="d18af912abacbf84219be37ab3b42a9abcf10d2a" }
 byte-slice-cast = { version = "1.2.1", default-features = false }
 rlp = { version = "0.5", default-features = false }
-hex-literal = { version = "0.3.1", optional = true }
+hex-literal = { version = "0.3.4", optional = true }
 
 frame-benchmarking = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.30", default-features = false, optional = true }
 frame-support = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.30", default-features = false }
@@ -48,6 +48,7 @@ std = [
     "scale-info/std",
     "frame-support/std",
     "frame-system/std",
+    'frame-benchmarking?/std',
     "sp-core/std",
     "sp-io/std",
     "sp-runtime/std",
